[PaymentRequest] Move Payum command providers to PayumBundle
