+++ src/Sylius/Bundle/PayumBundle/PaymentRequest/CommandProvider/AuthorizeCommandProvider.php
@@ -11,21 +11,21 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\CoreBundle\PaymentRequest\CommandProvider\Payum;
+namespace Sylius\Bundle\PayumBundle\PaymentRequest\CommandProvider;
 
 use Sylius\Bundle\CoreBundle\PaymentRequest\CommandProvider\PaymentRequestCommandProviderInterface;
-use Sylius\Bundle\PayumBundle\PaymentRequest\Command\StatusPaymentRequest;
+use Sylius\Bundle\PayumBundle\PaymentRequest\Command\AuthorizePaymentRequest;
 use Sylius\Component\Payment\Model\PaymentRequestInterface;
 
-final class StatusCommandProvider implements PaymentRequestCommandProviderInterface
+final class AuthorizeCommandProvider implements PaymentRequestCommandProviderInterface
 {
     public function supports(PaymentRequestInterface $paymentRequest): bool
     {
-        return $paymentRequest->getAction() === PaymentRequestInterface::ACTION_STATUS;
+        return $paymentRequest->getAction() === PaymentRequestInterface::ACTION_AUTHORIZE;
     }
 
     public function provide(PaymentRequestInterface $paymentRequest): object
     {
-        return new StatusPaymentRequest($paymentRequest->getHash()?->toBinary());
+        return new AuthorizePaymentRequest($paymentRequest->getHash()?->toBinary());
     }
 }
