+++ src/Sylius/Bundle/PayumBundle/Resources/config/services/payment_request/command_provider.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Sylius Sp. z o.o.
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<container
+    xmlns="http://symfony.com/schema/dic/services"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
+>
+    <services>
+        <defaults public="true" />
+
+        <service id="sylius.payum.payment_request.command_provider.offline" class="Sylius\Bundle\PaymentBundle\CommandProvider\ActionsCommandProvider">
+            <argument type="tagged_locator" tag="sylius.payum.payment_request.command_provider.offline" index-by="action" />
+            <tag name="sylius.payment_request.command_provider" gateway-factory="offline" priority="100" />
+        </service>
+
+        <service id="sylius.payum.payment_request.command_provider.offline.capture" class="Sylius\Bundle\PayumBundle\PaymentRequest\CommandProvider\CaptureCommandProvider">
+            <tag name="sylius.payum.payment_request.command_provider.offline" action="capture" />
+        </service>
+
+        <service id="sylius.payum.payment_request.command_provider.offline.status" class="Sylius\Bundle\PayumBundle\PaymentRequest\CommandProvider\StatusCommandProvider">
+            <tag name="sylius.payum.payment_request.command_provider.offline" action="status" />
+        </service>
+    </services>
+</container>
